+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -7,17 +7,6 @@
 
 import Foundation
 
-
-private class FeedCachePolicy {
-    private init() {}
-    private static let calender = Calendar(identifier: .gregorian)
-    private static var maxCacheAgeInDays: Int { return 7 }
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
-        guard let maxCacheAge = calender.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else { return false }
-        return date < maxCacheAge
-    }
-}
-
 public final class LocalFeedLoader {
     private let store: FeedStore
     private let currentDate: () -> Date
